Switch configuration executor List type to Sequence
The list of executors in configuration is intended to be read only.
Using Sequence for this makes that more explicit, as Sequences cannot
be modified.

This is part of a move towards using Sequence rather than List in
many places (although the runtime type remains a list) because Sequence
has better behaviour with subclasses of its parameter type: it is
covariant, rather than invariant.

This PR rearranges the validation logic a bit so that there is no
longer an arbitrary property setter for the executor property - it was
always a bug to replace the executor list in a config once parsl has
been configured, and removing this property setter enforces that more
strongly.